more strict rules about what may occur in a type
- to disallow various unorthodox ways of printing compilable c++.

commit also hides python like if scoping rules behind if expressions with a declared 'noscope' type (TODO fix a few existing bugs with this e.g. hoisting of assigns in if test expression)

commit also fixes a bug with 'noscope' disabled where an assignment in
an 'if' test expression wasn't treated as a definition when determining if
a subsequent reassignment in the "then block" is an 'auto' definition.